configured weapons

+++ src/main/java/com/wixl/better/items/DaggerItem.java
@@ -20,9 +20,9 @@ public class DaggerItem extends ToolItem implements Vanishable, DualWield {
 	private final float attackDamage;
 	private final Multimap<EntityAttribute, EntityAttributeModifier> attributeModifiers;
 
-	public DaggerItem(ToolMaterial toolMaterial, int attackDamage, float attackSpeed, Item.Settings settings) {
+	public DaggerItem(ToolMaterial toolMaterial, float attackDamage, float attackSpeed, Item.Settings settings) {
 		super(toolMaterial, settings);
-		this.attackDamage = (float)attackDamage + toolMaterial.getAttackDamage();
+		this.attackDamage = attackDamage + toolMaterial.getAttackDamage();
 		ImmutableMultimap.Builder<EntityAttribute, EntityAttributeModifier> builder = ImmutableMultimap.builder();
 		builder.put(EntityAttributes.GENERIC_ATTACK_DAMAGE, new EntityAttributeModifier(ATTACK_DAMAGE_MODIFIER_ID, "Weapon modifier", (double)this.attackDamage, EntityAttributeModifier.Operation.ADDITION));
 		builder.put(EntityAttributes.GENERIC_ATTACK_SPEED, new EntityAttributeModifier(ATTACK_SPEED_MODIFIER_ID, "Weapon modifier", (double)attackSpeed, EntityAttributeModifier.Operation.ADDITION));
